feat: adding account creation and persistence
